+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.2] - 2020-07-17
 
-- Show artist name and track name with link to Jamendo website
+## Added
+- Show artist name and track name with link to Jamendo website when playing audio
+
+## Changed
+- Better alerts
+- Merged code for showing segments and trajectories
 
 ## [0.2.1] - 2020-05-29